% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{tc_connect}
\alias{tc_connect}
\title{Connect to a ClickHouse database}
\usage{
tc_connect(
  host = NULL,
  port = NULL,
  user = NULL,
  password = NULL,
  db = "development"
)
}
\arguments{
\item{host}{Character scalar. ClickHouse host name or IP address. Defaults
to the value of the \code{CLICKHOUSE_HOST} environment variable.}

\item{port}{Integer scalar. HTTP port. Defaults to \code{CLICKHOUSE_PORT}
(coerced to integer).}

\item{user}{Character scalar. Database user name. Defaults to
\code{CLICKHOUSE_USER}.}

\item{password}{Character scalar. Database password. Defaults to
\code{CLICKHOUSE_PASSWORD}.}

\item{db}{Character scalar. Database to connect to. Defaults to
\code{"development"}.}
}
\value{
A \code{\link[DBI]{DBIConnection}} object that can be passed to any
  \code{tc_fetch_*()} function.
}
\description{
Opens a DBI connection to a ClickHouse instance using the ClickHouseHTTP
driver. By default all credentials are read from environment variables so
that no secrets appear in source code.
}
\examples{
\dontrun{
  con <- tc_connect()
  con <- tc_connect(db = "production")
  DBI::dbDisconnect(con)
}

}
\seealso{
\code{\link{tc_fetch_old_erp}}, \code{\link{tc_fetch_new_erp}},
  \code{\link{tc_fetch_providers}}
}
